fix(common): int keys should encode like `abi.encode` (#1001)
* add signed keys test

* update signed int key encoding to work like abi.encode

* test more key encodings

* build example

* generate cli test tables

* build templates

* revert codegen clean up

+++ packages/cli/contracts/src/codegen/tables/Statics.sol
@@ -108,7 +108,7 @@ library Statics {
   ) internal view returns (uint256 v1) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -132,7 +132,7 @@ library Statics {
   ) internal view returns (uint256 v1) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -147,7 +147,7 @@ library Statics {
   function setV1(uint256 k1, int32 k2, bytes16 k3, address k4, bool k5, Enum1 k6, Enum2 k7, uint256 v1) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -171,7 +171,7 @@ library Statics {
   ) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -193,7 +193,7 @@ library Statics {
   ) internal view returns (int32 v2) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -217,7 +217,7 @@ library Statics {
   ) internal view returns (int32 v2) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -232,7 +232,7 @@ library Statics {
   function setV2(uint256 k1, int32 k2, bytes16 k3, address k4, bool k5, Enum1 k6, Enum2 k7, int32 v2) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -256,7 +256,7 @@ library Statics {
   ) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -278,7 +278,7 @@ library Statics {
   ) internal view returns (bytes16 v3) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -302,7 +302,7 @@ library Statics {
   ) internal view returns (bytes16 v3) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -317,7 +317,7 @@ library Statics {
   function setV3(uint256 k1, int32 k2, bytes16 k3, address k4, bool k5, Enum1 k6, Enum2 k7, bytes16 v3) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -341,7 +341,7 @@ library Statics {
   ) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -363,7 +363,7 @@ library Statics {
   ) internal view returns (address v4) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -387,7 +387,7 @@ library Statics {
   ) internal view returns (address v4) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -402,7 +402,7 @@ library Statics {
   function setV4(uint256 k1, int32 k2, bytes16 k3, address k4, bool k5, Enum1 k6, Enum2 k7, address v4) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -426,7 +426,7 @@ library Statics {
   ) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -448,7 +448,7 @@ library Statics {
   ) internal view returns (bool v5) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -472,7 +472,7 @@ library Statics {
   ) internal view returns (bool v5) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -487,7 +487,7 @@ library Statics {
   function setV5(uint256 k1, int32 k2, bytes16 k3, address k4, bool k5, Enum1 k6, Enum2 k7, bool v5) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -511,7 +511,7 @@ library Statics {
   ) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -533,7 +533,7 @@ library Statics {
   ) internal view returns (Enum1 v6) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -557,7 +557,7 @@ library Statics {
   ) internal view returns (Enum1 v6) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -572,7 +572,7 @@ library Statics {
   function setV6(uint256 k1, int32 k2, bytes16 k3, address k4, bool k5, Enum1 k6, Enum2 k7, Enum1 v6) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -596,7 +596,7 @@ library Statics {
   ) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -618,7 +618,7 @@ library Statics {
   ) internal view returns (Enum2 v7) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -642,7 +642,7 @@ library Statics {
   ) internal view returns (Enum2 v7) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -657,7 +657,7 @@ library Statics {
   function setV7(uint256 k1, int32 k2, bytes16 k3, address k4, bool k5, Enum1 k6, Enum2 k7, Enum2 v7) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -681,7 +681,7 @@ library Statics {
   ) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -703,7 +703,7 @@ library Statics {
   ) internal view returns (StaticsData memory _table) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -727,7 +727,7 @@ library Statics {
   ) internal view returns (StaticsData memory _table) {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -759,7 +759,7 @@ library Statics {
 
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -791,7 +791,7 @@ library Statics {
 
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -888,7 +888,7 @@ library Statics {
   ) internal pure returns (bytes32[] memory _keyTuple) {
     _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -900,7 +900,7 @@ library Statics {
   function deleteRecord(uint256 k1, int32 k2, bytes16 k3, address k4, bool k5, Enum1 k6, Enum2 k7) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
@@ -923,7 +923,7 @@ library Statics {
   ) internal {
     bytes32[] memory _keyTuple = new bytes32[](7);
     _keyTuple[0] = bytes32(uint256((k1)));
-    _keyTuple[1] = bytes32(uint256(uint32((k2))));
+    _keyTuple[1] = bytes32(uint256(int256((k2))));
     _keyTuple[2] = bytes32((k3));
     _keyTuple[3] = bytes32(uint256(uint160((k4))));
     _keyTuple[4] = _boolToBytes32((k5));
